scripts/checksyscalls.sh: Add usage message, fix -Wno-error
Currently 'scripts/checksyscalls.sh' does not
provide guidance when executed without specifying
a compiler, instead it attempts to execute
'-Wno-error' parameter as a command,
resulting in 'Wno-error: Command not found' error.

This patch adds a usage message that is displayed
when no compiler is provided as parameter to
improve clarity.

Signed-off-by: Ritvik Gupta <ritvikfoss@gmail.com>